requirements.txt
absl-py==0.9.0
apex==0.1
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
bounded-pool-executor==0.0.3
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
CIRpy==1.0.2
Click==7.0
control==0.8.3
cssselect2==0.2.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
fluids==0.1.78
future==0.18.2
google-auth==1.19.1
google-auth-oauthlib==0.4.1
googledrivedownloader==0.4
grpcio==1.30.0
h5py==2.10.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.2.0
jupyter-core==4.6.3
kiwisolver==1.1.0
line-profiler==3.0.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mordred==1.2.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.42
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.9.0
plyfile==0.7.1
pqdm==0.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycpd==2.0.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch-lightning==0.8.5
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.6
QtPy==1.9.0
rdflib==4.2.2
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.13.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
terminado==0.8.3
testpath==0.4.4
thermo==0.1.40
tifffile==2020.9.3
tinycss2==1.0.2
torch==1.4.0
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.4.0
torch-sparse==0.4.3
torch-spline-conv==1.1.1
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
xgboost==1.2.1
xlrd==1.2.0
zipp==3.1.0